Implemented cart; Some refactoring; Updated readme; Redone books filtering.

+++ utils/link_tinymce_langs.sh
@@ -1,2 +1,5 @@
 #!/usr/bin/env bash
-cd static/libs/tinymce && ln -s ../tinymce-i18n/langs langs
+if [ ! -e static/libs/tinymce/langs ]
+then
+    cd static/libs/tinymce && ln -s ../tinymce-i18n/langs langs
+fi
